+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 
 - Fixed component not install when push modal
 - Updated select drawer menu when undo and redo form
+- Fixed component orientation `RTL` not update when show form
 
 ## [1.2.0] - 2024-09-21